refactor: move mock client to _test.go file
